Add more computational geometry algorithms
Reconstruct Repository structure by language, like Java or web.

Add two Convex Hull algorithms: Brute force and Graham's Scan, as well as visualization.
Add visualization to Bentley Ottmann's algrithom, that is, Geometric Intersection.
Reconstruct the visualization program in the triangluation program, make it aligned to the one taking advantage of normalization in graphics, which is the one other programs use as well. And there is also a web-based teaching program for the triangulation algorithm, which is more interactive than Java implementation.
Add Point Location ( Trapezoid Map and Search structure ), as well as visualization.